+++ src/frameworks/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod privileges;
 pub mod signal_handler;
 pub mod upstream;